# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_kiarashAlizadeh_herotech_internal_dto.CreateItemRequest:
    properties:
      list_price:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
...
  github_com_kiarashAlizadeh_herotech_internal_dto.ItemResponse:
    properties:
      base_price:
        type: integer
      created_at:
        type: string
      id:
        type: string
      list_price:
        type: integer
      name:
        type: string
      owner_id:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  github_com_kiarashAlizadeh_herotech_internal_dto.ListItemRequest:
    properties:
      list_price:
        type: integer
    type: object
...
  github_com_kiarashAlizadeh_herotech_internal_dto.PaginatedResponse-github_com_kiarashAlizadeh_herotech_internal_dto_ItemResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.ItemResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.PaginationMeta'
    type: object
  github_com_kiarashAlizadeh_herotech_internal_dto.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@Aethoria.com
    name: API Support
  description: Core infrastructure backend for managing Guild Wallets, Item Trading,
    and Legendary Auctions.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DRAGON MARKET PLACE API
  version: "1.0"
paths:
...
  /items:
    get:
      description: Fetches currently unpurchased inventory orders with optional filtering
        by class tier type
      parameters:
      - description: Filter by type (common, rare, legendary)
        in: query
        name: type
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.PaginatedResponse-github_com_kiarashAlizadeh_herotech_internal_dto_ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all available catalog items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Mints a dynamic asset directly into the guild's private inventory
        vault ('sold' state initially)
      parameters:
      - description: Owner Guild UUID
        in: header
        name: X-Guild-ID
        required: true
        type: string
      - description: Item metadata specs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.CreateItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mint a new inventory item
      tags:
      - items
  /items/{id}:
    get:
      description: Retrieves individual asset evaluation logs by its primary identifier
        key
      parameters:
      - description: Item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get item profile details
      tags:
      - items
  /items/{id}/buy:
    post:
      description: Triggers instant direct transaction fulfillment loop logic for
        Common or Rare asset categories
      parameters:
      - description: Buyer Guild UUID
        in: header
        name: X-Guild-ID
        required: true
        type: string
      - description: Item UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Purchase item directly (Limit Order)
      tags:
      - items
  /items/{id}/list:
    post:
      consumes:
      - application/json
      description: Exposes a non-legendary inventory item to the public market catalog
        with a fixed price
      parameters:
      - description: Owner Guild UUID
        in: header
        name: X-Guild-ID
        required: true
        type: string
      - description: Item UUID
        in: path
        name: id
        required: true
        type: string
      - description: Market listing pricing
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.ListItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List an item on the marketplace
      tags:
      - items
swagger: "2.0"